Tentativa de implemtar dtos, exclusão dos controllers api, a aplicação inicia, porem muita coisa parou de funcionar

+++ src/main/java/com/sistema/clinica/services/PacienteService.java
@@ -4,7 +4,10 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 
+import com.sistema.clinica.models.dtos.PacienteDTO;
+import com.sistema.clinica.models.dtos.mappers.PacienteMapper;
 import com.sistema.clinica.models.enums.Role;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -31,57 +34,77 @@ public class PacienteService {
 
 
 
-    public List<Paciente> findAll(){
-        return pacienteRepository.findAll();
+    public List<PacienteDTO> findAll(){
+        //o stream decompõe a list, a função map transforma paciente em pacienteDTO e o collector coloca tudo na lista novamente
+        return pacienteRepository.findAll().
+                stream().
+                map(PacienteMapper::toDTO).
+                collect(Collectors.toList());
     }
 
-    public Paciente findById(Long id){
-        Optional<Paciente> obj = pacienteRepository.findById(id);
-        return obj.orElseThrow(()-> new ResourceAccessException("Paciente não encontrado"));
+    public PacienteDTO findById(Long id){
+        Paciente paciente = pacienteRepository.findById(id)
+                .orElseThrow(()-> new ResourceAccessException("Paciente não encontrado"));
+        return PacienteMapper.toDTO(paciente);
     }
 
-    public Paciente insert(Paciente obj){
+    public PacienteDTO insert(PacienteDTO dto){
+        Paciente paciente = PacienteMapper.toEntity(dto);
 
-        String senhaCriptografada = passwordEncoder.encode(obj.getPassword());
-        obj.setPassword(senhaCriptografada);
+        if(dto.password() != null && !dto.password().isEmpty()){
+            paciente.setPassword(passwordEncoder.encode(dto.password()));
+        }else {
+            throw new IllegalArgumentException("Senha é obrigatória para cadastro de pacientes");
+        }
 
-        if (obj.getUsername() == null || obj.getUsername().isEmpty()) {
-            obj.setUsername(obj.getEmail());
+        if (paciente.getUsername() == null || paciente.getUsername().isEmpty()) {
+            paciente.setUsername(paciente.getEmail());
         }
 
         Set<Role> roles = EnumSet.of(Role.ROLE_PACIENTE);  // Define 'USER' como a role padrão
-        obj.setRoles(roles);
-        return pacienteRepository.save(obj);
+        paciente.setRoles(roles);
+
+        Paciente pacienteSalvo = pacienteRepository.save(paciente);
+
+        return PacienteMapper.toDTO(pacienteSalvo);
     }
 
     public void delete(Long id){
+
+        if (!pacienteRepository.existsById(id)) {
+            throw new ResourceAccessException("Paciente com ID " + id + " não encontrado.");
+        }
         pacienteRepository.deleteById(id);
     }
 
-    public Paciente update(Long id, Paciente obj){
+
+    public PacienteDTO update(Long id, PacienteDTO dto){
         try{
             Paciente entity = pacienteRepository
-                    .findById(id).orElseThrow(() -> new EntityNotFoundException("Paciente não encontrado com matrícula: " + id));
-            updateData(entity, obj);
-            return pacienteRepository.save(entity);
+                    .findById(id)
+                    .orElseThrow(() -> new EntityNotFoundException("Paciente não encontrado com matrícula: " + id));
+            updateDataFromDTO(entity, dto);
+            Paciente pacienteAtualizado = pacienteRepository.save(entity);
+            return PacienteMapper.toDTO(pacienteAtualizado);
         } catch (Exception e) {
             throw new RuntimeException("Erro ao atualizar o Paciente: " + e.getMessage(), e);
         }
     }
 
-    private void updateData(Paciente entity, Paciente obj) {
-        entity.setNome(obj.getNome());
-        entity.setCpf(obj.getCpf());
-        entity.setEmail(obj.getEmail());
-        entity.setTelefone(obj.getTelefone());
-        if (obj.getPassword() != null && !obj.getPassword().isBlank()) {
-            if (!obj.getPassword().startsWith("$2a$")) {
-                entity.setPassword(passwordEncoder.encode(obj.getPassword()));
+    private void updateDataFromDTO(Paciente entity, PacienteDTO dto) {
+        entity.setNome(dto.nome());
+        entity.setCpf(dto.cpf());
+        entity.setEmail(dto.email());
+        entity.setTelefone(dto.telefone());
+        if (dto.password() != null && !dto.password().isBlank()) {
+            if (!dto.password().startsWith("$2a$")) {
+                entity.setPassword(passwordEncoder.encode(dto.password()));
             }
         }
 
 
     }
+
 
 
 }
